backend/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.7.4
python-engineio==3.11.2
python-socketio==4.4.0
six==1.14.0
typing-extensions==3.7.4.1
yarl==1.4.2
